# projecteBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conecta Salud API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Llistar pacients",
                "description": "Retorna tots els pacients amb opcions de filtratge per zona, operador i estat",
                "operationId": "0df0d65f049161931654c3e188ca8f93",
                "parameters": [
                    {
                        "name": "zoneId",
                        "in": "query",
                        "description": "ID de la zona",
                        "required": false
                    },
                    {
                        "name": "operatorId",
                        "in": "query",
                        "description": "ID de l'operador assignat",
                        "required": false
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Estat del pacient",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pacients recuperats correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PatientResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error recuperant els pacients"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Crear pacient",
                "description": "Crea un nou pacient amb idiomes i persones de contacte",
                "operationId": "2149729f1bf1af8877352f04be3a683f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pacient creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la creació del pacient"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/{patient}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Mostrar pacient",
                "description": "Mostra els detalls d'un pacient concret",
                "operationId": "1aee48d18288527e7444c4af98753647",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "description": "ID del pacient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pacient recuperat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pacient no trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Actualitzar pacient",
                "description": "Actualitza un pacient existent i la seva informació relacionada",
                "operationId": "35a55f91c065962f3ab252b99c5ce17e",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "description": "ID del pacient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pacient actualitzat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error actualitzant el pacient"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Patients"
                ],
                "summary": "Eliminar pacient",
                "description": "Elimina un pacient específic",
                "operationId": "4d33007c68e93fcbc05bdbb7da29d89e",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "description": "ID del pacient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pacient eliminat correctament"
                    },
                    "400": {
                        "description": "Error eliminant el pacient"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/{id}/calls": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Historial de trucades per pacient",
                "description": "Recupera totes les trucades associades a un pacient",
                "operationId": "ddb952213ecea2d6aff23799b20eac97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pacient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial de trucades recuperat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CallResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pacient no trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePatientRequest": {
                "description": "Validació per a l'emmagatzematge de pacients",
                "required": [
                    "fullName",
                    "birthDate",
                    "fullAddress",
                    "dni",
                    "healthCardNumber",
                    "phone",
                    "email",
                    "zoneId",
                    "operatorId",
                    "languages"
                ],
                "properties": {
                    "fullName": {
                        "description": "Nom complet del pacient",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Joan Pérez"
                    },
                    "birthDate": {
                        "description": "Data de naixement del pacient",
                        "type": "string",
                        "format": "date",
                        "example": "1980-01-01"
                    },
                    "fullAddress": {
                        "description": "Adreça completa del pacient",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Carrer de l'Example, 123, 08001 Barcelona"
                    },
                    "dni": {
                        "description": "DNI del pacient",
                        "type": "string",
                        "maxLength": 20,
                        "example": "12345678A"
                    },
                    "healthCardNumber": {
                        "description": "Número de targeta sanitària del pacient",
                        "type": "string",
                        "maxLength": 20,
                        "example": "12345678901234567890"
                    },
                    "phone": {
                        "description": "Telèfon del pacient",
                        "type": "string",
                        "maxLength": 15,
                        "example": "600123456"
                    },
                    "email": {
                        "description": "Correu electrònic únic del pacient",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "joan.perez@example.com"
                    },
                    "zoneId": {
                        "description": "Identificador de la zona",
                        "type": "integer",
                        "example": 1
                    },
                    "operatorId": {
                        "description": "Identificador de l'operador",
                        "type": "integer",
                        "example": 1
                    },
                    "languages": {
                        "type": "array",
                        "items": {
                            "description": "Idiomes parlats pel pacient",
                            "type": "string",
                            "example": "ca"
                        }
                    },
                    "status": {
                        "description": "Estat del pacient",
                        "type": "string",
                        "example": "admitted"
                    },
                    "contactPersons": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "firstName": {
                                    "description": "Nom de la persona de contacte",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Maria"
                                },
                                "lastName": {
                                    "description": "Cognom de la persona de contacte",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "García"
                                },
                                "phone": {
                                    "description": "Telèfon de la persona de contacte",
                                    "type": "string",
                                    "maxLength": 15,
                                    "example": "600123456"
                                },
                                "relationship": {
                                    "description": "Relació amb el pacient",
                                    "type": "string",
                                    "example": "parent"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdatePatientRequest": {
                "description": "Validation for updating a patient",
                "required": [
                    "fullName",
                    "birthDate",
                    "fullAddress",
                    "dni",
                    "healthCardNumber",
                    "phone",
                    "email",
                    "zoneId",
                    "operatorId",
                    "languages"
                ],
                "properties": {
                    "fullName": {
                        "description": "Full name of the patient",
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "birthDate": {
                        "description": "Birth date of the patient",
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "fullAddress": {
                        "description": "Full address of the patient",
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Main St, Anytown, USA"
                    },
                    "dni": {
                        "description": "DNI of the patient",
                        "type": "string",
                        "maxLength": 20,
                        "example": "12345678A"
                    },
                    "healthCardNumber": {
                        "description": "Health card number of the patient",
                        "type": "string",
                        "maxLength": 20,
                        "example": "HC123456789"
                    },
                    "phone": {
                        "description": "Phone number of the patient",
                        "type": "string",
                        "maxLength": 15,
                        "example": "+1234567890"
                    },
                    "email": {
                        "description": "Unique email of the patient",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "john.doe@example.com"
                    },
                    "zoneId": {
                        "description": "Zone ID where the patient resides",
                        "type": "integer",
                        "example": 1
                    },
                    "personalFamilySituation": {
                        "description": "Personal family situation of the patient",
                        "type": "string",
                        "example": "Single"
                    },
                    "healthSituation": {
                        "description": "Health situation of the patient",
                        "type": "string",
                        "example": "Good"
                    },
                    "housingSituation": {
                        "description": "Housing situation of the patient",
                        "type": "string",
                        "example": "Own house"
                    },
                    "personalAutonomy": {
                        "description": "Personal autonomy of the patient",
                        "type": "string",
                        "example": "Independent"
                    },
                    "economicSituation": {
                        "description": "Economic situation of the patient",
                        "type": "string",
                        "example": "Stable"
                    },
                    "operatorId": {
                        "description": "ID of the operator handling the patient",
                        "type": "integer",
                        "example": 2
                    },
                    "languages": {
                        "type": "array",
                        "items": {
                            "description": "Languages spoken by the patient",
                            "type": "string",
                            "example": "en"
                        }
                    },
                    "status": {
                        "description": "Status of the patient",
                        "type": "string",
                        "example": "admitted"
                    },
                    "contactPersons": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "firstName": {
                                    "description": "First name of the contact person",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Jane"
                                },
                                "lastName": {
                                    "description": "Last name of the contact person",
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Doe"
                                },
                                "phone": {
                                    "description": "Phone number of the contact person",
                                    "type": "string",
                                    "maxLength": 15,
                                    "example": "+1234567890"
                                },
                                "relationship": {
                                    "description": "Relationship of the contact person to the patient",
                                    "type": "string",
                                    "example": "Sibling"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "LanguageResource": {
                "description": "Esquema del recurs Language",
                "properties": {
                    "id": {
                        "description": "Identificador de l'idioma",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'idioma",
                        "type": "string",
                        "example": "Català"
                    }
                },
                "type": "object"
            },
...
            "PatientResource": {
                "description": "Esquema del recurs Patient",
                "properties": {
                    "id": {
                        "description": "Identificador del pacient",
                        "type": "integer",
                        "example": 1
                    },
                    "fullName": {
                        "description": "Nom complet del pacient",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "birthDate": {
                        "description": "Data de naixement del pacient",
                        "type": "string",
                        "format": "date",
                        "example": "1980-01-01"
                    },
                    "fullAddress": {
                        "description": "Adreça completa del pacient",
                        "type": "string",
                        "example": "123 Main St, City, Country"
                    },
                    "dni": {
                        "description": "DNI del pacient",
                        "type": "string",
                        "example": "12345678A"
                    },
                    "healthCardNumber": {
                        "description": "Número de targeta sanitària del pacient",
                        "type": "string",
                        "example": "HC123456789"
                    },
                    "phone": {
                        "description": "Telèfon del pacient",
                        "type": "string",
                        "example": "+123456789"
                    },
                    "email": {
                        "description": "Correu electrònic del pacient",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "personalFamilySituation": {
                        "description": "Situació familiar personal del pacient",
                        "type": "string",
                        "example": "Single"
                    },
                    "healthSituation": {
                        "description": "Situació de salut del pacient",
                        "type": "string",
                        "example": "Healthy"
                    },
                    "housingSituation": {
                        "description": "Situació d'habitatge del pacient",
                        "type": "string",
                        "example": "Own house"
                    },
                    "personalAutonomy": {
                        "description": "Autonomia personal del pacient",
                        "type": "string",
                        "example": "Independent"
                    },
                    "economicSituation": {
                        "description": "Situació econòmica del pacient",
                        "type": "string",
                        "example": "Stable"
                    },
                    "operatorId": {
                        "description": "Identificador de l'operador",
                        "type": "integer",
                        "example": 2
                    },
                    "languages": {
                        "description": "Llengües parlades pel pacient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LanguageResource"
                        }
                    },
                    "contactPersons": {
                        "description": "Persones de contacte del pacient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContactPersonResource"
                        }
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "operator": {
                        "$ref": "#/components/schemas/OperatorResource"
                    },
                    "calls": {
                        "description": "Trucades associades al pacient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CallResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}